Bugfix for microVM shutdown, bump versions of all packages (0.2.1-beta for core and 1.0.0-rc.2 for hosts)

+++ FirecrackerSharp/Boot/UnrestrictedVm.cs
@@ -52,7 +52,10 @@ public class UnrestrictedVm : Vm
 
     protected override void CleanupAfterShutdown()
     {
-        IHostFilesystem.Current.DeleteFile(SocketPath!);
+        if (IHostFilesystem.Current.FileOrDirectoryExists(SocketPath!))
+        {
+            IHostFilesystem.Current.DeleteFile(SocketPath!);
+        }
     }
 
     /// <summary>
